requirements.txt
# Core ML and Data Processing
pandas==2.0.3
numpy>=1.26.0
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.2

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP Client for testing
httpx==0.25.2
requests==2.31.0

# Monitoring and Metrics
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.19.0

# Configuration and Logging
pyyaml==6.0.1
python-json-logger==2.0.7

# Visualization and Analysis
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Jupyter and Development
jupyter==1.0.0
ipykernel==6.25.2
notebook==7.0.6

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development Tools
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1

# Security and Validation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Environment management
python-dotenv==1.0.0
